Split peer supervisor CLI commands
