+++ loopx/cli_commands/support_control_registry.py
@@ -1,0 +1,68 @@
+from __future__ import annotations
+
+from pathlib import Path
+
+from ..history import load_registry
+from ..paths import DEFAULT_RUNTIME_ROOT, global_registry_path
+from ..registry import registry_goals, resolve_state_file
+
+
+def default_public_scan_root() -> str:
+    return str(Path(__file__).resolve().parents[2])
+
+
+def fallback_global_registry(registry_path: Path, runtime_root_arg: str | None) -> Path:
+    if registry_path.exists():
+        return registry_path
+    runtime_root = Path(runtime_root_arg).expanduser() if runtime_root_arg else DEFAULT_RUNTIME_ROOT
+    fallback_registry = global_registry_path(runtime_root)
+    return fallback_registry if fallback_registry.exists() else registry_path
+
+
+def explicit_global_registry(runtime_root_arg: str | None) -> Path:
+    runtime_root = Path(runtime_root_arg).expanduser() if runtime_root_arg else DEFAULT_RUNTIME_ROOT
+    return global_registry_path(runtime_root)
+
+
+def resolve_heartbeat_active_state(
+    *,
+    goal_id: str,
+    active_state_arg: str | None,
+    registry_path: Path,
+    runtime_root_arg: str | None,
+    allow_global_goal_lookup_fallback: bool = True,
+) -> tuple[Path | None, Path | None, str]:
+    if active_state_arg:
+        active_state = Path(active_state_arg).expanduser()
+        return active_state, active_state, "explicit"
+
+    resolved_registry = fallback_global_registry(registry_path, runtime_root_arg)
+    registry = load_registry(resolved_registry)
+    goal = next((item for item in registry_goals(registry) if item.get("id") == goal_id), None)
+    if goal is None and allow_global_goal_lookup_fallback:
+        global_registry = explicit_global_registry(runtime_root_arg)
+        if global_registry != resolved_registry and global_registry.exists():
+            global_payload = load_registry(global_registry)
+            global_goal = next(
+                (item for item in registry_goals(global_payload) if item.get("id") == goal_id),
+                None,
+            )
+            if global_goal is not None:
+                resolved_registry = global_registry
+                registry = global_payload
+                goal = global_goal
+    if goal is None:
+        raise ValueError(
+            f"goal_id not found in registry for heartbeat active-state lookup: {goal_id}"
+        )
+    repo_text = str(goal.get("repo") or "")
+    if not repo_text:
+        raise ValueError(f"{goal_id}: registry goal has no repo for active-state lookup")
+    state_file = resolve_state_file(Path(repo_text).expanduser(), goal.get("state_file"))
+    if state_file is None:
+        raise ValueError(f"{goal_id}: registry goal has no state_file for active-state lookup")
+    if not state_file.exists():
+        raise FileNotFoundError(
+            f"{goal_id}: registry-declared active state file does not exist: {state_file}"
+        )
+    return None, state_file, f"registry:{resolved_registry}"
