+++ loopx/cli_commands/todo_argument_validation.py
@@ -95,3 +95,24 @@ def validate_shared_todo_options(args: argparse.Namespace) -> None:
             "`todo add/update --role user --task-class user_gate`; --from, --limit, and "
             "--trigger are only supported by `todo suggest`"
         )
+
+
+def validate_capability_gap_options(args: argparse.Namespace) -> None:
+    if not args.capability_gap_status:
+        return
+    if args.todo_command not in {"add", "update"}:
+        raise ValueError("--capability-gap-status is supported only by todo add/update")
+    if args.role != "agent":
+        raise ValueError("--capability-gap-status requires --role agent")
+    if not args.target_capabilities:
+        raise ValueError(
+            "--capability-gap-status requires at least one --target-capability"
+        )
+    if (
+        args.capability_gap_status in {"fixed", "real_callsite_verified"}
+        and not args.evidence
+    ):
+        raise ValueError(
+            "fixed and real_callsite_verified capability gaps require "
+            "public-safe --evidence"
+        )
